feat(dev): profile renders with React Scan
Add react-scan as a dev dependency and start it from the renderer entry.
React Scan gives an always-on overlay (live FPS, re-render highlighting,
a slowdown profiler, and an LLM-prompt export) that is far better suited to
diagnosing the high-frequency re-renders of long agent loops than a manual
profiler.

It must patch React's DevTools hook before react-dom's reconciler
initializes, so the entry is split: main.tsx starts React Scan first (dev
only), then dynamically imports and mounts the app from app.tsx. The dev
branch is tree-shaken from the production bundle.
